Fix broken stories in Storybook.
